+++ services/core/java/core-flow-service/src/main/java/org/ubiquia/core/flow/service/factory/NodeFactory.java
@@ -51,7 +51,7 @@ public class NodeFactory {
 
         logger.info("...completed building node {} for graph {}...",
             nodeEntity.getName(),
-            nodeEntity.getGraph().getName());
+            nodeEntity.getParentGraph().getName());
         return node;
     }