+++ services/core/java/core-flow-service/src/test/java/org/ubiquia/core/flow/service/cluster/synchronization/entity/FlowEventSynchronizationServiceTest.java
@@ -119,7 +119,7 @@ public class FlowEventSynchronizationServiceTest {
         this.testHelper.registerAndDeploy(ontology, graph);
 
         var nodeEntity = this.nodeRepository
-            .findByGraphNameAndName(graph.getName(), node.getName())
+            .findByParentGraphNameAndName(graph.getName(), node.getName())
             .orElseThrow();
 
         var nodeDto = new Node();